Refactor and consolidate utility functions; remove unused files and update FolderNode interface to include pageData
