+++ packages/studiocms_core/src/sdk-utils/StudioCMSSDK.ts
@@ -109,7 +109,13 @@ export class StudioCMSSDK {
 		const folderMap: Record<string, FolderNode> = {};
 
 		for (const folder of folders) {
-			folderMap[folder.id] = { id: folder.id, name: folder.name, page: false, children: [] };
+			folderMap[folder.id] = {
+				id: folder.id,
+				name: folder.name,
+				pageData: null,
+				page: false,
+				children: [],
+			};
 		}
 
 		// Build the tree
