+++ packages/studiocms_dashboard/assets/astro.d.ts
@@ -1,1 +1,10 @@
 /// <reference path="../../../playground/.astro/types.d.ts" />
+
+interface Window {
+	theme: {
+		setTheme: (theme: 'system' | 'dark' | 'light') => void;
+		getTheme: () => 'system' | 'dark' | 'light';
+		getSystemTheme: () => 'light' | 'dark';
+		getDefaultTheme: () => 'system' | 'dark' | 'light';
+	};
+}
