+++ packages/studiocms_dashboard/assets/components/ThemeManager.astro
@@ -1,0 +1,57 @@
+---
+---
+
+<script is:inline>
+    window.theme ??= (() => {
+        const defaultTheme = "system";
+        const storageKey = "studiocms-theme-preference";
+        const store =
+            typeof localStorage !== "undefined"
+            ? localStorage
+            : { getItem: () => null, setItem: () => {} };
+    
+        const mediaMatcher = window.matchMedia("(prefers-color-scheme: light)");
+        let systemTheme = mediaMatcher.matches ? "light" : "dark";
+        mediaMatcher.addEventListener("change", (event) => {
+            systemTheme = event.matches ? "light" : "dark";
+            applyTheme(theme.getTheme());
+        });
+    
+        function applyTheme(theme) {
+            const resolvedTheme = theme === "system" ? systemTheme : theme;
+            document.documentElement.dataset.theme = resolvedTheme;
+            document.documentElement.style.colorScheme = resolvedTheme;
+            document.dispatchEvent(
+            new CustomEvent("theme-changed", {
+                detail: { theme, systemTheme, defaultTheme },
+            })
+            );
+        }
+    
+        function setTheme(theme = defaultTheme) {
+            store.setItem(storageKey, theme);
+            applyTheme(theme);
+        }
+    
+        function getTheme() {
+            return store.getItem(storageKey) || defaultTheme;
+        }
+    
+        function getSystemTheme() {
+            return systemTheme;
+        }
+    
+        function getDefaultTheme() {
+            return defaultTheme;
+        }
+    
+        return { setTheme, getTheme, getSystemTheme, getDefaultTheme };
+    })();
+    theme.setTheme(theme.getTheme());
+</script>
+
+<script>
+    document.addEventListener("astro:after-swap", () =>
+        window.theme.setTheme(window.theme.getTheme()),
+    );
+</script>
