+++ packages/studiocms_dashboard/assets/components/islands/configuration/LightVsDark.astro
@@ -1,4 +1,5 @@
 ---
+import { Image } from 'astro:assets';
 
 interface Props {
 	light: ImageMetadata;
@@ -9,9 +10,10 @@ const { light, dark } = Astro.props;
 ---
 <div class="wrapper">
     <div class="before">
-      <img class="content-image light-image-holder" src={light.src} draggable="false"/>   </div>
+      <Image width={640} height={360} class="content-image light-image-holder" src={light} draggable="false" alt='Light Preview' loading={'eager'} />   
+    </div>
     <div class="after">
-      <img class="content-image dark-image-holder" src={dark.src} draggable="false"/>
+      <Image width={640} height={360} class="content-image dark-image-holder" src={dark} draggable="false" alt='Dark Preview' loading={'eager'} />
     </div>
 </div>
