+++ packages/studiocms_dashboard/assets/components/islands/content-mgmt/CreatePage.astro
@@ -13,7 +13,7 @@ import {
 <div id="create-page-container" data-link={StudioCMSRoutes.sdk.pages}>
     <form id="page-create-form">
 
-      <Tabs syncKey='edit-page-tabs' variant={'starlight'}>
+      <Tabs variant={'starlight'}>
         <TabItem label='Basic Information' icon={'information-circle-20-solid'} color={'primary'}>
           <div class="tab-section">
             <div class="form-row">
@@ -61,9 +61,9 @@ import {
           <TabItem label='Page Content' icon={'document-text-20-solid'} color={'success'}>
             <div class="tab-section">
               <div class="page-content-editor">
-                  <div id="editor-toolbar"></div>
+                  <div id="create_editor-toolbar"></div>
                   <div class="editor-container">
-                      <textarea id="page-content" name="page-content"></textarea>
+                      <textarea id="create_page-content" name="page-content"></textarea>
                   </div>
               </div>
             </div>
@@ -167,14 +167,14 @@ import {
 
         // Setup the TinyMDE editor
         // See https://github.com/jefago/tiny-markdown-editor/tree/main for more details
-        const pageContentTextarea = createPageForm.querySelector('#page-content');
-        const editorToolbar = createPageForm.querySelector('#editor-toolbar');
+        const pageContentTextarea = createPageForm.querySelector('#create_page-content');
+        const editorToolbar = createPageForm.querySelector('#create_editor-toolbar');
         // @ts-expect-error - Added TinyMDE to the window object
-        var tinyMDE = new TinyMDE.Editor({ textarea: pageContentTextarea });
+        var createTinyMDE = new TinyMDE.Editor({ textarea: pageContentTextarea });
         // @ts-expect-error - Added TinyMDE to the window object
         var commandBar = new TinyMDE.CommandBar({
           element: editorToolbar,
-          editor: tinyMDE,
+          editor: createTinyMDE,
           commands: [
             {
               name: 'bold',
@@ -263,12 +263,12 @@ import {
         });
 
         // @ts-expect-error - Added TinyMDE to the window object
-        tinyMDE.addEventListener('change', e => { 
+        createTinyMDE.addEventListener('change', e => { 
           pageContentTextarea.textContent = e.content; 
         });
 
         // @ts-expect-error - Added TinyMDE to the window object
-        tinyMDE.addEventListener("drop", function (event) {
+        createTinyMDE.addEventListener("drop", function (event) {
           let formData = new FormData();
 
           // You can add use event.dataTransfer.items or event.dataTransfer.files
