+++ packages/studiocms_dashboard/assets/components/islands/content-mgmt/EditPage.astro
@@ -25,7 +25,7 @@ function getParentFolderValue(value: string | null): string | undefined {
 <div id="edit-page-container" data-link={StudioCMSRoutes.sdk.pages}>
     <form id="edit-page-form">
 
-        <Tabs syncKey='edit-page-tabs' variant={'starlight'}>
+        <Tabs variant={'starlight'}>
           <TabItem label='Basic Information' icon={'information-circle-20-solid'} color={'primary'}>
             <div class="tab-section">
                 <div class="form-row">
