+++ packages/studiocms_dashboard/assets/routes/configuration.astro
@@ -17,13 +17,13 @@ const lang = 'en-us';
 
   <div slot="header">
     <PageHeader title="Site Configuration">
-      <Button variant="solid" color="primary" type="submit" form="site-config-form">
-        <Icon slot="start-content" name="check" width={24} height={24} />
+      <Button variant="solid" color="primary" size="sm" type="submit" form="site-config-form">
+        <Icon slot="start-content" name="check-20-solid" width={20} height={20} />
         Save Changes
       </Button>
     </PageHeader>
   </div>
   
-  <ConfigForm transition:persist transition:persist-props/>
+  <ConfigForm />
 
 </Layout>
