--- packages/studiocms_robotstxt/src/utils/index.ts
@@ -1,1 +1,0 @@
-export { default as measureExecutionTime } from './measureExecutionTime.js';
