--- packages/studiocms_robotstxt/src/utils/measureExecutionTime.ts
@@ -1,7 +1,0 @@
-export default function measureExecutionTime(callback: () => void): number {
-	const startTime = performance.now();
-	callback();
-	const endTime = performance.now();
-	const executionTime = Math.floor(endTime - startTime);
-	return executionTime;
-}
